Assignment2/requirements.txt
# Core scientific stack
numpy>=1.22,<1.27
scipy>=1.9,<1.13
pandas>=1.4,<2.2

# Torch & vision
torch>=1.12,<2.2
torchvision>=0.13,<0.16
torchmetrics>=0.10,<1.4
timm>=0.6.12,<0.10.0

# Object detection / COCO
pycocotools>=2.0.6

# Utility libraries
tqdm>=4.62
pyyaml>=5.1
pyparsing>=2.3.1
decorator>=4.0.2,<6.0
imageio>=2.5,<3.0
requests>=2.26
pillow>=9.2,<11.0  # compatible with torchvision & moviepy
pip install albumentations
